FEP-61cf: Describe the OpenWebAuth protocol
I recently built a prototype OpenWebAuth implementation which is able
to interoperate with Hubzilla. This involved reverse engineering the
protocol and making some notes on how it works.

When I was done I realised that my notes could easily be polished into
a first draft of a specification. This change contains the second
iteration of this polishing process.

The intention is that it should be possible to write an interoperable
implementation from this document alone, without the need for any
reverse engineering.