FEP-61cf: More info on webfinger and signatures
Added a description of the "http://purl.org/openwebauth/v1#redirect"
webfinger rel value, and a suggestion that new implementations should
use it instead of a hard-coded value.

Added a reference to the recent "ActivityPub and HTTP Signatures"
SocialCG report - but kept the description of how OpenWebAuth diverges
slightly from this by using the Authorization header.